requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
daphne==4.1.2
defusedxml==0.7.1
Django==4.2
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
fastapi==0.115.0
filelock==3.16.1
Flask==3.0.3
greenlet==3.1.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
outcome==1.3.0.post0
packaging==24.1
parsel==1.9.1
prompt_toolkit==3.0.48
Protego==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
queuelib==1.7.0
redis==5.1.1
requests==2.32.3
requests-file==2.1.0
rich==13.8.1
Scrapy==2.11.2
scrapy-splash==0.9.0
selenium==4.25.0
service-identity==24.1.0
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
starlette==0.38.6
tldextract==5.1.2
tomli==2.0.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
zope.interface==7.0.3
